Added Manager for routed processes
